updated two files
i found some bugs in the smart contract, couldnt vote for candidate 5, BillieE, and the voting start wasnt defined nor duration in the correct manner. added those to the constructor. correct all code for those problems and now the duration of the election is set for 11 mins for the DEMO